Extract VertexToolControllerParts
