+++ lib/TbUiLib/include/ui/VertexToolController.h
@@ -28,6 +28,9 @@ namespace tb::ui
 class Tool;
 class VertexTool;
 
+template <typename T>
+class VertexToolPartBase;
+
 class VertexToolController : public VertexToolControllerBase<VertexTool>
 {
 protected:
